pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"


[project]
name = "tidypolars4sci_ext"
version = "0.1.0"
description = "Extension package for tidypolars4sci — adds extra dplyr/tidyr verbs, joins, ranking, set operations, and more."
authors = [{name = "Diogo Ferrari", email = "diogoferrari@gmail.com"}]
maintainers = [{name = "Diogo Ferrari", email = "diogoferrari@gmail.com"}]
license = "MIT"
license-files = ["LICENSE"]
readme = "README.md"
requires-python = ">=3.9"
keywords = ["polars", "tidyverse", "dataframe", "tidy data", "extension"]
classifiers = [
  "Development Status :: 4 - Beta",
  "Intended Audience :: Science/Research",
  "Topic :: Scientific/Engineering"
]
dependencies = [
    "tidypolars4sci>=0.0.1.22",
]

[tool.hatch.build.targets.wheel]
packages = ["tidypolars4sci_ext"]

[project.urls]
Homepage = "https://github.com/mdmanurung/tidypolars4sci"
Documentation = "https://github.com/mdmanurung/tidypolars4sci"
Issues = "https://github.com/mdmanurung/tidypolars4sci/issues/"